refactor: move method to compatibility_layers and update legendary method
