+++ src/backend/storeManagers/hyperplay/games.ts
@@ -9,9 +9,7 @@ import {
   LicenseConfigValidateResult,
   ChannelReleaseMeta,
   SiweValues,
-  UpdateArgs,
-  WineCommandArgs,
-  Runner
+  UpdateArgs
 } from '../../../common/types'
 import { hpLibraryStore } from './electronStore'
 import { sendFrontendMessage, getMainWindow } from 'backend/main_window'
@@ -74,11 +72,10 @@ import { cancelQueueExtraction } from 'backend/downloadmanager/downloadqueue'
 import { captureException } from '@sentry/electron'
 import Store from 'electron-store'
 import i18next from 'i18next'
-import { gameManagerMap } from '..'
-import { runWineCommand } from 'backend/launcher'
 import { DEV_PORTAL_URL } from 'common/constants'
 import getPartitionCookies from 'backend/utils/get_partition_cookies'
 import { prepareBaseGameForModding } from 'backend/ipcHandlers/mods'
+import { runWineCommandOnGame } from 'backend/utils/compatibility_layers'
 
 interface ProgressDownloadingItem {
   DownloadItem: DownloadItem
@@ -311,32 +308,6 @@ export async function importGame(
   rmSync(path.join(pathName, `${appName}.json`))
   writeManifestFile(appName, gameInLibrary.install)
   return { stderr: '', stdout: '' }
-}
-
-export async function runWineCommandOnGame(
-  runner: Runner,
-  appName: string,
-  { commandParts, wait = false, protonVerb, startFolder }: WineCommandArgs
-): Promise<ExecResult> {
-  if (gameManagerMap[runner].isNative(appName)) {
-    logError(
-      `runWineCommand called on native game: ${appName}`,
-      LogPrefix.Backend
-    )
-    return { stdout: '', stderr: '' }
-  }
-  const { folder_name, install } = gameManagerMap[runner].getGameInfo(appName)
-  const gameSettings = await gameManagerMap[runner].getSettings(appName)
-
-  return runWineCommand({
-    gameSettings,
-    installFolderName: folder_name,
-    gameInstallPath: install.install_path,
-    commandParts,
-    wait,
-    protonVerb,
-    startFolder
-  })
 }
 
 type DistArgs = {
